+++ api/Startup.cs
@@ -13,7 +13,6 @@ using dotenv.net;
 
 using static api.Common.UserRoles;
 using api.Models;
-using api.Models.FilterParser;
 using api.Services;
 
 namespace api
@@ -106,7 +105,7 @@ namespace api
 			});
 			#endregion
 
-			services.AddTransient<IParser, Parser>();
+			services.AddTransient<IFilterParser, FilterParser>();
 
 			services.AddMvc();
 		}
